minor improvements

+++ cgo.h
@@ -16,7 +16,7 @@ int luaGo_call(lua_State *L, lua_CFunction f);
 
 int luaGo_load(lua_State *L, void *data, const char *chunkname, const char *mode);
 
-void luaGo_pushGoValue(lua_State *L, unsigned long ud);
+void luaGo_pushGoValue(lua_State *L, uintptr_t ud);
 
 int luaGo_callGoFunction(lua_State *L);
 int luaGo_getGoFunction(lua_State *L, int idx);
